+++ hvega/src/Graphics/Vega/VegaLite.hs
@@ -134,6 +134,7 @@ module Graphics.Vega.VegaLite
        , VL.GradientStops
        , VL.Opacity
        , VL.SelectionLabel
+       , VL.StyleLabel
        , VL.VegaExpr
        , VL.ZIndex
        , VL.toHtml
@@ -1065,6 +1066,14 @@ import qualified Graphics.Vega.VegaLite.Transform as VL
 -- constructor has been added to the 'VL.Channel' type, and 'VL.RNumberLists'
 -- (Vega-Lite 4.4) to 'VL.ScaleRange'.
 --
+-- Named styles have been added for axes as well as marks. As mentioned below,
+-- this involves deprecating the previous constructors for naming styles,
+-- as there are now separate configuration options: 'VL.AxisNamedStyles'
+-- and 'VL.MarkNamedStyles'. The 'VL.AStyle' and 'VL.AxStyle' options have been
+-- added to 'VL.AxisConfig' and 'VL.AxisProperty' respectively.
+-- The `VL.StyleLabel` type alias has been added to help the documentation, but
+-- provides no extra type safety.
+--
 -- __Breaking Change__
 --
 -- The 'VL.ConcatStyle' and 'VL.FacetStyle' constructors for
@@ -1130,20 +1139,33 @@ import qualified Graphics.Vega.VegaLite.Transform as VL
 -- with the addition of 'VL.LeSymbolLimit', 'VL.LeTickCount', 'VL.LeTitleLineHeight',
 -- and 'VL.LeUnselectedOpacity' constructors.
 --
+-- The axis configuration and property types ('VL.AxisConfig' and 'VL.AxisProperty')
+-- have gained the Vega-Lite 4.4 'VL.LabelOffset' and 'VL.AxLabelOffset' constructors.
+-- Note that version 4.4.0 of the Vega-Lite specification has these fields
+-- as strings but @hvega@ outputs numbers, as the specification was wrong
+-- and will be fixed in the next version.
+--
 -- 'VL.ConfigurationProperty' has added new constructors:
 -- 'VL.AxisQuantitative', 'VL.AxisTemporal', 'VL.BoxplotStyle',
 -- 'VL.ErrorBandStyle', 'VL.ErrorBarStyle', 'VL.FontStyle' (Vega-Lite 4.3),
 -- 'VL.HeaderColumnStyle', 'VL.HeaderFacetStyle', 'VL.HeaderRowStyle',
 -- 'VL.ImageStyle', and 'VL.RepeatStyle'.
 --
--- The @Autosize@, @Background@, @CountTitle@, @FieldTitle@,
--- @Legend@, @NumberFormat@, @Padding@, @Projection@, @Range@,
--- @Scale@, @TimeFormat@, and @View@ constructors for
--- 'VL.ConfigurationProperty' are now deprecated, and are replaced by
--- 'VL.AutosizeStyle', 'VL.BackgroundStyle', 'VL.CountTitleStyle',
--- 'VL.FieldTitleStyle', 'VL.LegendStyle', 'VL.NumberFormatStyle',
--- 'VL.PaddingStyle', 'VL.ProjectionStyle', 'VL.RangeStyle',
--- 'VL.ScaleStyle', 'VL.TimeFormatStyle', and 'VL.ViewStyle' respectively.
+-- __Deprecated symbols__:
+--
+-- 'VL.ConfigurationProperty' has seen a large number of deprecations,
+-- as a number of constructors have been renamed:
+--
+-- * @NamedStyle@ and @NamedStyles@ have been replaced by 'VL.MarkNamedStyles';
+--
+-- * @Autosize@, @Background@, @CountTitle@, @FieldTitle@,
+--   @Legend@, @NumberFormat@, @Padding@, @Projection@, @Range@,
+--   @Scale@, @TimeFormat@, and @View@
+--   constructors have been replaced by
+--   'VL.AutosizeStyle', 'VL.BackgroundStyle', 'VL.CountTitleStyle',
+--   'VL.FieldTitleStyle', 'VL.LegendStyle', 'VL.NumberFormatStyle',
+--   'VL.PaddingStyle', 'VL.ProjectionStyle', 'VL.RangeStyle',
+--   'VL.ScaleStyle', 'VL.TimeFormatStyle', and 'VL.ViewStyle' respectively.
 
 -- $update0500
 -- The @0.5.0.0@ release now creates specifications using version 4
@@ -1563,7 +1585,8 @@ import qualified Graphics.Vega.VegaLite.Transform as VL
 --
 -- The 'VL.ConfigurationProperty' type, used with 'VL.configuration', has
 -- gained 'VL.ConcatStyle', 'VL.FacetStyle', 'VL.GeoshapeStyle', 'VL.HeaderStyle',
--- 'VL.NamedStyles', and 'VL.TrailStyle' constructors.
+-- @NamedStyles@ (renamed to 'VL.MarkNamedStyles' in 0.6.0.0),
+-- and 'VL.TrailStyle' constructors.
 --
 -- The 'VL.ConcatConfig' type was added for use with the 'VL.ConcatStyle',
 -- and the 'VL.FacetConfig' type for the 'VL.FacetStyle'